weekly-report: Start work on parameterization
It is now possible to apply constraints from the command line
to generate a personal weekly report:

  weekly-report assignee = daniel.thompson@linaro.org

The constraint can be any fragment of JQL but the @ character it
treated differently to a normal JQL query (it is automatically
escaped to simplify commandl ine use meaning it cannot play
a syntactic role in a query).

Also included here are some hacks to help me extract status from
my KWG cards.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>