glimpse: Improve robustness of worklog extraction
Currently a worklog with insufficient vertical whitespace may lose
the entries following the titles. For example:
~~~
h3. Progress
This line will go missing because it is spuriously linked to the title.
This line will also go missing.

This line is OK because it gets picked up by the new paragraph detection.
~~~

This patch fixes things so we will never lose the lines following the
title regardless of what whitespace is used. However the new paragraph
detect still relies on double line break (or bullets) so all lines
following the title will be reported by glimpse as a single bullet
point.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>