glance: interact: Exploit composable commands
Now that fetch is seperated out we can use it directly from the interact
command to automatically fetch data if we don't have something to
interact with.

This is especially useful for interact because it cannot be used from a
pipe (redirecting the stdin makes it very hard to interact with the data).